requirements.txt
azure-iot-device==2.12.0
pyserial==3.5
